This seems to work well as of 2/16
This hasn't been tested with the 120VAC to 12VDC power supply involved but it does fix the problem of turning the dust collector off in the event that a saw is turned on during the countdown sequence. It now checks the values again by brute force before turning it off.